Delete mysterious notebook
it doesn't appear in my local version of master, but i'm pretty sure i moved it or a version of it into the notebooks directory. i've never in all my years seen a file in the repo that refuses to appear locally; i have no idea how this is even possible.